+++ docs/source/changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.10.4 (2025-07-14)
+- Make pydot optional dependency for visualization of flow execution
+- Make psycopg2/adbc-driver-postgres optional dependency
+
 ## 0.10.3 (2025-07-10)
 - Fix: Do not implicitly depend on `kazoo`
 - Fix some misleading warnings